Add confirmation views
A "forgot password" button is added to the login page that takes the user to `/forgotPassword`.
`/confirmation` and `/resetPassword` are added and only accept a non null token in the query params.
Custom routes are added to App for these new pages. They currently display the nav/sidebar which is a side effect that will need to be addressed in a future build.
Files are reorganized into folders.